Update logic to correctly track jobs with multiple dependencies
Also:
Simplify pruning logic to be less error-prone
Fix if an input hash did not match it would not check if the output exists or not
Publish to NPM as v1.1.1

+++ index.js
@@ -43,7 +43,8 @@ module.exports = function gbcache(cache_opts, task_opts) {
   let index;
   let new_index;
 
-  function recordToFilename2(record, file_entry) {
+  function recordToFilename2(file_entry) {
+    assert(file_entry.output_hash);
     let fn = safeFilename(file_entry.relative);
     let ext_idx = fn.lastIndexOf('.');
     let ext = '';
@@ -51,7 +52,19 @@ module.exports = function gbcache(cache_opts, task_opts) {
       ext = fn.slice(ext_idx);
       fn = fn.slice(0, ext_idx);
     }
-    return path.join(cache_folder, `${fn}#${record.input_hash.slice(-7)}${ext}`);
+    return `${cache_folder}/${fn}#${file_entry.output_hash.slice(-7)}${ext}`;
+  }
+
+  function inputHash(record) {
+    let { inputs } = record;
+    assert(inputs[0].relative === '&the');
+    let ret = [];
+    ret.push(inputs[0].input_hash);
+    for (let ii = 1; ii < inputs.length; ++ii) {
+      let elem = inputs[ii];
+      ret.push(elem.bucket, elem.relative, elem.input_hash);
+    }
+    return ret.join(':');
   }
 
   let cache_hit;
@@ -68,23 +81,36 @@ module.exports = function gbcache(cache_opts, task_opts) {
             return;
           }
           line = line.split(';');
-          let [relative, ver, input_hash, ...outputs] = line;
+          let [relative, ver, deps_string, ...outputs] = line;
           let entry = index[relative] = {
             relative,
             ver: Number(ver),
-            input_hash,
-            files: [],
+            inputs: [],
+            outputs: [],
           };
+          let parts = deps_string.split(':');
+          assert(parts.length % 3 === 1); // should be 1 + 3xN
+          entry.inputs.push({
+            relative: '&the',
+            input_hash: parts[0],
+          });
+          for (let ii = 1; ii < parts.length; ii+=3) {
+            entry.inputs.push({
+              bucket: parts[ii],
+              relative: parts[ii+1],
+              input_hash: parts[ii+2],
+            });
+          }
           if (outputs.length % 2) {
             // odd number, first entry must be identity name
-            entry.files.push({
+            entry.outputs.push({
               relative,
               output_hash: outputs[0],
             });
             outputs.splice(0, 1);
           }
           for (let ii = 0; ii < outputs.length; ii += 2) {
-            entry.files.push({
+            entry.outputs.push({
               relative: outputs[ii],
               output_hash: outputs[ii+1],
             });
@@ -102,22 +128,58 @@ module.exports = function gbcache(cache_opts, task_opts) {
     let file = job.getFile();
 
     assert(!file.relative.includes(';')); // not supported, used internally as a delimiter
+    assert(!file.relative.includes(':')); // not supported, used internally as a delimiter
 
-    let input_hash = hash(file.contents);
     function cacheMiss() {
       // run job, save output to cache if needed
       ++cache_miss;
       task_opts.func(job, function (err) {
-        if (!err) {
-          let new_entry;
-          if (do_cache_write) {
-            new_entry = new_index[file.relative] = {
-              relative: file.relative,
-              ver: version,
-              input_hash,
-              files: [],
-            };
+        if (err) {
+          return void done(err);
+        }
+        let deps = job.getDeps();
+        assert(deps.includes(`${file.bucket}:${file.relative}`)); // must have at least the main file!
+
+        let new_entry;
+        if (do_cache_write) {
+          new_entry = new_index[file.relative] = {
+            relative: file.relative,
+            ver: version,
+            inputs: [],
+            outputs: [],
+          };
+        }
+        function gatherDeps(next) {
+          if (!do_cache_write) {
+            return void next();
           }
+          asyncEach(deps, function (dep_name, next) {
+            if (dep_name === `${file.bucket}:${file.relative}`) {
+              new_entry.inputs.push({
+                relative: '&the',
+                input_hash: hash(file.contents),
+              });
+              return void next();
+            }
+            job.depAdd(dep_name, function (err, depfile) {
+              let input_hash;
+              if (err) {
+                input_hash = 'null';
+              } else {
+                assert(depfile.contents instanceof Buffer);
+                input_hash = hash(depfile.contents);
+              }
+              let [bucket, relative] = dep_name.split(':');
+              new_entry.inputs.push({
+                bucket,
+                relative,
+                input_hash,
+              });
+              next();
+            });
+          }, next);
+        }
+        function hashOutputs() {
           let output_queue = job.getOutputQueue();
           let output_keys = Object.keys(output_queue);
           for (let ii = 0; ii < output_keys.length; ++ii) {
@@ -127,7 +189,7 @@ module.exports = function gbcache(cache_opts, task_opts) {
             assert(Buffer.isBuffer(contents));
             if (do_cache_write) {
               let output_hash = hash(contents);
-              new_entry.files.push({
+              new_entry.outputs.push({
                 relative,
                 output_hash,
                 needs_write: true,
@@ -136,20 +198,50 @@ module.exports = function gbcache(cache_opts, task_opts) {
             }
           }
         }
-        done(err);
+        gatherDeps(function () {
+          hashOutputs();
+          done();
+        });
       });
     }
 
     let cache_record = index[file.relative];
-    if (cache_record &&
-      cache_record.ver === version &&
-      cache_record.input_hash === input_hash
-    ) {
-      // cache appears valid
+    if (!cache_record) {
+      return void cacheMiss();
+    }
+    // check all deps
+    function checkDeps(next) {
+      let deps_match = true;
+      asyncEach(cache_record.inputs, function (elem, next) {
+        if (elem.relative === '&the') {
+          let new_hash = hash(file.contents);
+          if (new_hash !== elem.input_hash) {
+            deps_match = false;
+          }
+          return void next();
+        }
+        job.depAdd(`${elem.bucket}:${elem.relative}`, function (err, depfile) {
+          let new_hash;
+          if (err) {
+            new_hash = 'null';
+          } else {
+            assert(depfile.contents instanceof Buffer);
+            new_hash = hash(depfile.contents);
+          }
+          if (new_hash !== elem.input_hash) {
+            deps_match = false;
+          }
+          next();
+        });
+      }, function () {
+        next(deps_match);
+      });
+    }
+    // check all outputs - note: need to do this even if we have a cache miss, so we know which files to re-write
+    function checkOutputs(force_miss) {
       let outputs = [];
-      let force_miss = false;
-      asyncEach(cache_record.files, function (file_entry, next, idx) {
-        let cached_file = recordToFilename2(cache_record, file_entry);
+      asyncEach(cache_record.outputs, function (file_entry, next, idx) {
+        let cached_file = recordToFilename2(file_entry);
         fs.readFile(cached_file, function (err, buffer) {
           if (err) {
             job.warn(`gbcache: unable to load file referenced by cache: ${cached_file} (${err})`);
@@ -182,9 +274,11 @@ module.exports = function gbcache(cache_opts, task_opts) {
         cache_record.seen = true;
         done();
       });
-    } else {
-      cacheMiss();
     }
+    checkDeps(function (deps_match) {
+      let force_miss = cache_record.ver !== version || !deps_match;
+      checkOutputs(force_miss);
+    });
   }
   function finish() {
     if (cache_hit || cache_miss) {
@@ -194,20 +288,16 @@ module.exports = function gbcache(cache_opts, task_opts) {
     if (Object.keys(new_index).length || do_cache_rebuild) {
       // assemble new index, write it out
       gb.info(`  gbcache(${key}): Updating cache...`);
-      let to_prune = [];
       let unchanged = 0;
       for (let relative in index) {
         let old_entry = index[relative];
         let new_entry = new_index[relative];
-        let prune_old = false;
         if (!new_entry) {
           // either an old file not in the input set, or a file that was not changed
           if (old_entry.ver !== version) {
             // must not exist anymore, otherwise we would have had a job run
-            prune_old = true;
           } else if (do_cache_rebuild && !old_entry.seen) {
             // Forcing a rebuild and we did not get a job for this entry, must have been pruned from input data
-            prune_old = true;
           } else {
             // must assume this is just an input file that has not changed - place it in the new index
             new_index[relative] = old_entry;
@@ -215,40 +305,28 @@ module.exports = function gbcache(cache_opts, task_opts) {
           }
         } else {
           // have a new and old entry
-          if (old_entry.input_hash === new_entry.input_hash) {
-            // nothing changed (except perhaps the version), use the old data if appropriate
-            let all_unchanged = true;
-            for (let ii = 0; ii < old_entry.files.length; ++ii) {
-              let old_file = old_entry.files[ii];
-              // Look for same file in new_entry
-              let found = false;
-              for (let jj = 0; jj < new_entry.files.length; ++jj) {
-                let new_file = new_entry.files[jj];
-                if (new_file.relative === old_file.relative) {
-                  if (new_file.output_hash === old_file.output_hash) {
-                    // same, use it
-                    delete new_file.contents;
-                    delete new_file.needs_write;
-                  } // else: changed, write new contents, don't "prune" as the cache filename is unchanged
-                  found = true;
-                }
-              }
-              if (!found) {
-                all_unchanged = false;
-                to_prune.push(recordToFilename2(old_entry, old_file));
+          let all_unchanged = true;
+          for (let ii = 0; ii < old_entry.outputs.length; ++ii) {
+            let old_file = old_entry.outputs[ii];
+            // Look for same file in new_entry
+            let found = false;
+            for (let jj = 0; jj < new_entry.outputs.length; ++jj) {
+              let new_file = new_entry.outputs[jj];
+              if (new_file.relative === old_file.relative) {
+                if (new_file.output_hash === old_file.output_hash) {
+                  // same, use it
+                  delete new_file.contents;
+                  delete new_file.needs_write;
+                } // else: changed, write new contents, don't "prune" as the cache filename is unchanged
+                found = true;
               }
             }
-            if (all_unchanged) {
-              ++unchanged;
+            if (!found) {
+              all_unchanged = false;
             }
-          } else {
-            // new entry, new data, will write new one to disk, prune old files
-            prune_old = true;
           }
-        }
-        if (prune_old) {
-          for (let ii = 0; ii < old_entry.files.length; ++ii) {
-            to_prune.push(recordToFilename2(old_entry, old_entry.files[ii]));
+          if (all_unchanged) {
+            ++unchanged;
           }
         }
       }
@@ -260,16 +338,18 @@ module.exports = function gbcache(cache_opts, task_opts) {
         fs.mkdirSync(cache_folder);
       }
       // Write new files
-      let new_files = 0;
+      let new_outputs = 0;
+      let all_outputs = {};
       for (let relative in new_index) {
         let new_entry = new_index[relative];
-        for (let ii = 0; ii < new_entry.files.length; ++ii) {
-          let new_file = new_entry.files[ii];
+        for (let ii = 0; ii < new_entry.outputs.length; ++ii) {
+          let new_file = new_entry.outputs[ii];
+          let record_name = recordToFilename2(new_file);
+          all_outputs[record_name] = true;
           if (new_file.needs_write) {
-            let record_name = recordToFilename2(new_entry, new_file);
             gb.debug(`  gbcache(${key}): writing ${record_name}`);
             fs.writeFileSync(record_name, new_file.contents);
-            ++new_files;
+            ++new_outputs;
             delete new_file.needs_write;
             delete new_file.contents;
           }
@@ -280,9 +360,9 @@ module.exports = function gbcache(cache_opts, task_opts) {
       keys.sort();
       let index_txt = keys.map((relative) => {
         let record = new_index[relative];
-        let line = [relative, version, record.input_hash];
-        for (let ii = 0; ii < record.files.length; ++ii) {
-          let file_entry = record.files[ii];
+        let line = [relative, version, inputHash(record)];
+        for (let ii = 0; ii < record.outputs.length; ++ii) {
+          let file_entry = record.outputs[ii];
           if (file_entry.relative === record.relative) {
             line.splice(3, 0, file_entry.output_hash);
           } else {
@@ -294,21 +374,24 @@ module.exports = function gbcache(cache_opts, task_opts) {
       fs.writeFileSync(cache_index_path, index_txt);
 
       // Finally, prune old files
+      let existing_files = fs.readdirSync(cache_folder);
       let pruned = 0;
-      for (let ii = 0; ii < to_prune.length; ++ii) {
-        let filename = to_prune[ii];
-        gb.debug(`  gbcache(${key}): pruning ${filename}`);
-        if (fs.existsSync(filename)) {
-          try {
-            fs.unlinkSync(filename);
-            ++pruned;
-          } catch (err) {
-            gb.warn(`  gbcache(${key}): error deleting ${filename} (${err})`);
+      for (let ii = 0; ii < existing_files.length; ++ii) {
+        let filename = `${cache_folder}/${existing_files[ii]}`;
+        if (!all_outputs[filename]) {
+          gb.debug(`  gbcache(${key}): pruning ${filename}`);
+          if (fs.existsSync(filename)) {
+            try {
+              fs.unlinkSync(filename);
+              ++pruned;
+            } catch (err) {
+              gb.warn(`  gbcache(${key}): error deleting ${filename} (${err})`);
+            }
           }
         }
       }
 
-      gb.info(`  gbcache(${key}): ${new_files} new, ${pruned} pruned, ${unchanged} unchanged`);
+      gb.info(`  gbcache(${key}): ${new_outputs} new, ${pruned} pruned, ${unchanged} unchanged`);
 
       // TODO: add command line option where we force a version change, so as to
       //   force all files to be processed, so that we can prune files which no
